requirements.txt
networkx
tensorflow_gpu
tensorflow_probability_gpu
numpy
graph_nets
dm-sonnet